feat: format and push stac collection to github TDE-756 (#132)
If the target destination for `publish-copy` is `s3://linz-imagery/`
then the `collection.json` file will not be copied directly to S3 but
will have the root link to the catalog inserted, prettier formatting
applied, and it will be pushed to the linz/imagery GitHub repository,
creating a Pull Request.

![image](https://github.com/linz/topo-workflows/assets/21299036/4219c0d4-75d8-4e75-93fb-24858ed830f7)

If the target destination is **not** `s3://linz-imagery/` all files will
be copied to the target and no push to GitHub will be made.

![image](https://github.com/linz/topo-workflows/assets/21299036/88db1bb9-ae1c-4cfd-a969-6cc73e48dda9)