feat: Add semaphore to limit concurrent basemaps import workflows BM-822 (#157)
#### Description

Adds a semaphore to limit how many basemaps import workflows can run at
a time.

#### Intention

As we are about to bulk import the georeferenced historical imagery
surveys, it would be beneficial to add a semaphore to limit how many of
these can be run at once.

The limit of 10 is entirely arbitrary: very happy to change it to
whatever it seen as a better value before merging.

#### Checklist
- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title