refactor: publish-copy & publish-odr TDE-868 (#173)
**Why:**
Currently the two workflows: `publish-odr` and `publish-copy` are almost
direct copies of each other with a lot of duplicate code. This is a risk
for future work as the same code will need to be updated in two
different places, and it is likely to be forgotten/mistakes made.

**Description:**
`copy`, `create-manifest` and `publish-to-github` have been pulled out
into template workflow-templates that are each called by the tasks in
`publish-copy` and `publish-odr`

The alternative option is to remove `publish-odr` and add it as
options/parameters to `publish-copy`. I didn't do this as we had
discussed wanting to make sure the odr was separate and the future
possible enhancements of approval steps / restricting users who can run
the workflow.

---------

Co-authored-by: paulfouquet <86932794+paulfouquet@users.noreply.github.com>
Co-authored-by: Alice Fage <afage@linz.govt.nz>