fix: copy template bucket config ordering (#188)
The Argo Tasks copy command will register the first matching role, which
in the case of 'linz-imagery' was the read role, causing copying to
`linz-imagery` to fail.
It may be something that can be fixed in Argo Tasks or chunkd if there
is a robust way to work out the naming/logic of the roles.
I also removed the default values as we probably want to explicitly
specify the write roles in the upstream workflow.