ci: CDK cannot ask for deploy approval without terminal TDE-926 (#282)
#### Motivation

Sometimes `cdk deploy` asks for approval confirmation. Because using
`cdk deploy` part of the CI is not possible to interact with it, we need
to bypass this approval.

```
Deployment failed: Error: "--require-approval" is enabled and stack includes security-sensitive updates, but terminal (TTY) is not attached so we are unable to get a confirmation from the user
```

#### Modification

Using `--require-approval never` will bypass the confirmation.

#### Checklist

- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title