feat: Argo Workflows archive feature using RDS TDE-731 (#284)
#### Motivation

The aim of this change is to enable the Archived Workflows in Argo so
that older workflow runs can be referenced in the future e.g. to see
what was run when, what parameters were used.

#### Modification

This change provisions an AWS RDS PostgreSQL database and configures
Argo Workflows to use it, so that the Archived Workflows feature is
enabled.

#### Checklist

- [N/A] Tests updated
- [x] Docs updated
- [x] Issue linked in Title