feat: add ticket label to imagery related workflows TDE-956 (#304)
#### Motivation

A ticket (JIRA ticket) ID is a useful information to link to a workflow
for filtering purpose.

#### Modification

Ticket is passed as a parameter of the workflow to the
`linz.govt.nz/ticket` label value. If no label specified in the
parameter, its value is empty (which is fine according to [Kubernetes
documentation](https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#syntax-and-character-set))
but the label is still created.

#### Checklist

- [ ] Tests updated - N/A
- [x] Docs updated
- [x] Issue linked in Title