fix: create-thumbnails workflow should not try publishing data TDE-969 (#319)
#### Motivation

`create-thumbnails` workflow should only create the thumbnails in a
"temporary" place in the s3 `scratch` bucket. The files have to be
verified before copied over their final destination.

#### Modification

The WorkflowTemplate was containing publish copy tasks that should not
be done part of this workflow. Instead, a separate workflow to copy
files (`publish-copy`) should be run after someone verified the files.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated - N/A
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: Daniel Silk <dsilk@linz.govt.nz>