feat: add region label to imagery related workflows TDE-957, TDE-955 (#307)
#### Motivation

The region of the dataset is a useful information to be able to filter
on.

#### Modification

`region` is passed as a parameter of the workflow to the
`linz.govt.nz/region` label value. This is a mandatory field proposed as
an enum to the user. Any modification of the `region` list will have to
be applied at all the places it is used.
This change is also part of TDE-955 that will use this workflow
parameter as an input of a task/step to build the dataset title.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated - N/A
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: Alice Fage <afage@linz.govt.nz>