fix: change workflows to use new PAT TDE-980 (#326)
#### Motivation

Move to the new linz-li-bot-pat secret for GitHub pushes and remove
references to old secrets.

#### Modification

Remove the old secret and reference linz-li-bot PAT instead.
Remove the imagery-import workflow that was using the old github ssh key
as this workflow has been superseded by imagery-import-cogify.
Update the standardising-publish-import workflow that was still using
the old imagery-import workflow.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated - to be done by Basemaps team
https://toitutewhenua.atlassian.net/browse/BM-744
- [X] Issue linked in Title