feat: RDS usage metrics and alerts to Slack TDE-943 (#343)
#### Motivation

As a GDE developer, I want to retain CloudWatch logs and setup
monitoring alarms so that I can monitor the health, usage and cost of
the RDS instance and make sure that the Workflows Archive is reliable.

#### Modification

Add CloudWatch RDS metrics for available storage space and CPU
utilisation. Add alerts to an SNS Topic which the AWS ChatBot will relay
to the Argo dev alerts Slack channel.

The configuration for the Slack Channel is held in AWS Systems Manager
Parameter Store. The AWS ChatBot requires approval for the Slack Channel
(done via the AWS Console). This has been done.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated N/A
- [x] Docs updated (inline comments; not sure if more is needed).
- [X] Issue linked in Title