chore: update templates with Argo Tasks v3 (#450)
#### Motivation

Bring the templates into line with the workflows and workflowtemplates
to be consistent. It should have no functional effect as the upstream
workflows specify Argo Tasks `v3` which will override the templates, but
there is always the possibility that this might not be specified
upstream.

#### Modification

Change `version_argo_tasks` to `v3` in templates.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title