fix: regex for matching capture-area.geojson (#452)
#### Motivation

The previous regex looked for a match for the start of the filename i.e.
`^capture-area.geojson$`, however, when running on Argo Workflows, the
path given when matching is always the full S3 object (path), so this
did not match.

#### Modification

Change `^` to `/` to match the full filename. Also escape the full stop
to look for a literal full stop character.

#### Checklist

_If not applicable, provide explanation of why._

N/A workflow bug

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title