chore: remove unused test workflow (#456)
#### Motivation

This workflow does not seem to be used: it's pinned to a specific
version of argo-tasks, has a specific source. It looks like it is a
one-off workflow that has been pushed unintentionally.

#### Modification

Delete the unused test workflow. This can be bring back later if someone
needs it. Should be less specific and organize in directories.

#### Checklist

- [ ] Tests updated N/A
- [ ] Docs updated N/A
- [ ] Issue linked in Title N/A