feat: update publish-odr to remove copy and add new PR fields TDE-1100 TDE-1102 (#485)
#### Motivation

Provide more info for imagery and elevation PR reviews and copy the item
TIFFs and STAC files only after the PR is approved and merged.

#### Modification

The ticket and copy-option need to be part of the Pull Request.
The `create-manifest` and `copy` tasks are removed from the
`publish-odr` workflowtemplate.
The `publish-to-github` workflowtemplate is updated for the new fields
in Argo Tasks that are to be passed through to the Pull Request.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [x] Docs updated
- [x] Issue linked in Title