feat: add node local dns cache (#511)
#### Motivation

We are seeing a large amount of DNS failures with timeouts and
`EAI_AGAIN` responses, we believe it might be due to the DNS servers
getting overwhelmed when the cluster is scalling rapidly up.

By adding a DNS cache to every node there is a huge reduction of DNS
requests made over the wire to the primary coredns servers.


#### Modification

Adds a DNS cache to every node inside the cluster.

---------

Co-authored-by: paulfouquet <86932794+paulfouquet@users.noreply.github.com>