feat: default retryStrategy TDE-1112 (#506)
#### Motivation

We want every task within our workflows to retry (twice) automatically
and on a different node/host (network issues) when they fail.

#### Modification

Configure the `retryStrategy` at the `workflowDefaults` level.
Notes: 

- the `nodeAntiAffinity` to prioritize retrying on a different node/host
is not working due to an issue in the Argo Workflows system. A [PR is
opened](https://github.com/argoproj/argo-workflows/pull/12701) in the
Argo repo
- this change will make every task retrying on failure. Some of our
tasks (for example, `stac-validate`, `tileindex-validate`) are expected
to fail if the system invalidate something. They will still retry twice
in that case. This would be handle in a separate PR.

#### Checklist

- [ ] Tests updated NA
- [x] Docs updated 
- [x] Issue linked in Title