refactor: remove metadata.namespace from WorkflowTemplates and CronWorkflows TDE-1149 (#536)
#### Motivation

The `metadata.namespace` is overriden when the WorkflowTemplates and
CronWorkflows are deployed (using `--namespace`).
Having this data is redundant but also create issue while linting
locally (`argo lint --offline`) the WorkflowTemplates with
`templateRefs` when namespaces do not match.

#### Modification

Remove the default namespace in `metadata.namespace` from all
WorkflowTemplate and CronWorkflow files

#### Checklist

- [ ] Tests updated N/A
- [ ] Docs updated N/A
- [x] Issue linked in Title