ci: use --offline mode to lint workflows TDE-1149 TDE-1143 (#540)
#### Motivation

The current process of linting the workflows is complex as we need to
publish them on the kubernetes cluster. Using an offline mode makes
things simpler and quicker, removing the connection to the cluster.

#### Modification

Use the `--offline` mode to run `argo lint`, rather than linting from
the resources deployed on the cluster.

#### Checklist

- [ ] Tests updated
- [ ] Docs updated
- [X] Issue linked in Title