fix: node nodeAntiAffinity seems to break karpenter TDE-1112 BM-1018 (#545)
#### Motivation

We've noticed that karpenter is erroring since
https://github.com/linz/topo-workflows/pull/506. Errors makes us
thinking that it could be related. Since the retry on a new node using
`nodeAntiAffinity` [is not
working](https://github.com/argoproj/argo-workflows/pull/12701), it's a
good idea to remove it to see if it solved `karpenter` issues.

#### Modification

Remove the `affinity` in the `retryStrategy`.

#### Checklist

- [ ] Tests updated
- [x] Docs updated
- [x] Issue linked in Title