feat: include ticket number in basemaps pull requests TDE-1146 (#535)
#### Motivation

It is helpful to have all pull requests linked to the ticket, since we
now have the ticket number in the workflow we should pass it into
anything interacting with github

#### Modification

Adds ticket number to basemaps `create-pr` action.

This pull request needs https://github.com/linz/argo-tasks/pull/962 to
be merged first.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title