fix: stac-validate-collections retries should be managed by tpl-at-stac-validate only TDE-1178 (#585)
#### Motivation
In the case of a task in `tpl-at-stac-validate` fails more than twice
and the workflow came back failing, the entire
`stac-validate-collections` is retried twice which cause to run the
inner stac-validate workflow.

#### Modification

Avoid tasks in the `stac-validate-parallel` workflow to retry by
default.
Force `aws-list-collections` to retry twice if it fails.
This is a work around as I could not find a way to specify the
retryStrategy at a task level using a `templateRef`

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title