feat(basemaps): Support elevation config for the config import. BM-936 (#574)
#### Motivation

We are adding new config type for elevation data which leads to three
types of configs in Basemaps. So [pull
request](https://github.com/linz/argo-tasks/pull/953) will need to
select which config type that is needed for the imported imagery.

#### Modification

This is blocking by argo-tasks release. Not ready yet. Sorry to make
this as non draft PR, accidentally pushed enter when editing.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title