fix: retry only failed pods on standardising workflow TDE-1216 (#667)
#### Motivation

If the standardising workflow fails, the entire workflow is retried
twice, even the pods that succeeded. We only want the failed pods to
retry.

#### Modification

Specify no retries on the main template level, allowing the default
retry of 2 to remain on the individual steps.

Behaviour before this change:

![image](https://github.com/linz/topo-workflows/assets/21299036/fdb0feb7-d6da-4dfb-b331-0539e13a31b5)

Behaviour after this change:


![image](https://github.com/linz/topo-workflows/assets/21299036/4628ffe1-ae58-4f1e-b6a0-69812fee6468)


#### Checklist

- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title