feat: require a stac collection.json when importing into basemaps BM-1047 (#675)
#### Motivation

When importing imagery into basemaps sometimes URLs are misstyped and
import the wrong colleciton of imagery, by requiring a collection.json
to be present at the root of the imagery we can reducee the chance of
incorrect imagery being imported.

#### Modification

_Why is this change being made? What implications or other
considerations are there?_

Adds "--require-stac-collection" flag to the basemaps import process
which is defaulted to true.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title