fix: increase Fluent Bit Kubernetes Filter Buffer_Size to 128k TDE-1232 (#694)
#### Motivation

We're not seeing some logs propagated from Argo Workflows workflow pods
to Elasticsearch.
This issue and a solution can be found here:
https://github.com/argoproj/argo-workflows/issues/13314.

#### Modification

Increase the [Fluent Bit Kubernetes Filter
`Buffer_Size`](https://docs.fluentbit.io/manual/pipeline/filters/kubernetes#configuration-parameters)
from its default `32k` to `128k` as it seems to solve our problem.
If we experience this issue again, it could not be enough and we'll
proceed to another increase in time.

#### Checklist

- [ ] Tests updated
- [ ] Docs updated
- [X] Issue linked in Title