fix: entire workflow should not retry TDE-1236 (#708)
#### Motivation

With the current default `retryStrategy`, all templates are retried with
the limit put in place. When a pod withing a workflow reach its retry
limit without any success, the entrypoint template (`main`) of the
workflow is also retried using this `retryStrategy`. It is not efficient
as at that point the entire workflow is retried (and the failing
template/pod as well!).

#### Modification

- avoid the retry on entrypoint template for the workflows in
`workflows/`, not the `workflowTemplate` in `templates` as they have
only one template which is also their entrypoint.
- move the default `retryStrategy` to `templateDefaults` as it is there
that it belongs. This did not change the behaviour from the test I ran.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [x] Docs updated
- [x] Issue linked in Title