fix: resupplied Collection should keep its created datetime TDE-1340 (#885)
#### Motivation

When a dataset is resupplied (the dataset already exist / is already
published), the Collection should keep track of the updates. In order to
do that, the `created` datetime should remains the one that have been
introduced at the initial creation/publication of the dataset.

#### Modification

- pass the `odr_url` to the `collection_from_items.py` script in the
`create-collection` task, so the system retrieve the published
Collection to keep its information.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title