feat: add a workflowTemplate to create national DEM TDE-1166 (#890)
#### Motivation

Having a Nationally combined set of DEMs is useful for any data consumer
wanted to access to the entire country DEM. This dataset is aimed to be
updated on a regular basis to include all new single region/area dataset
that is newly published or updated.


![image](https://github.com/user-attachments/assets/e9d363f5-c586-4ced-8385-24d626ace894)


#### Modification

Add a WorkflowTemplate that create a nationally combined set of DEM
based off a configuration file that list the datasets to include in
order of priority. This WF template is able to be used to run a cron
workflow.

#### Checklist

- [ ] Tests updated
- [x] Docs updated
- [x] Issue linked in Title