feat: simple tiff merge script TDE-1370 (#950)
#### Motivation

The purpose of this script is to be able to merge the 1m hillshade
(covering part of the country) with the 8m hillshade (covering all of
the country) to get a hillshade product that covers the whole country in
the best available level of detail.
This workflow can be run standalone but is intended to be used as part
of a more complex workflow, e.g. one that creates a national hillshade
product from 1m DEM and 8m DEM.

#### Modification

Created new Workflow in `gdalmerge.yaml`.


#### Checklist

- [ ] Tests updated
- [x] Manual end to end tests
- [x] Docs updated
- [x] Issue linked in Title