feat: suspend national dem cron until the process works TDE-1359 (#973)
#### Motivation

The National DEM cron workflow is based on the fact that an existing
dataset can be updated without reprocessing the entire dataset (existing
files + new files). We don't want to remove this workflow but only
suspending it until the updating feature get released.
Suspending it on the deployed cron workflow has the limitation that when
a PR get merged to this repo, all the workflows get re-deployed,
removing the suspend flag...

#### Modification

Hardcode the suspend flag until the updating feature is released.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title