feat(basemaps): Topo raster maps standardisation workflow. BM-1128 (#932)
### Motivation

As a Basemaps user, I want to consume the NZTopo 50 & 250 Maps [as a
tile service].

### Modification

This work defines a new Argo `topo-maps-standardising` workflow and an
update to the existing `imagery-import-cogify` workflow. We have used
this work to generate and publish the [NZ Topo Gridless Maps] layer to
basemaps.

1. **`topo-maps-standardising` (new)**

This work defines a new Argo [topo-maps-standardising] workflow for
coordinating the standardisation of an NZTopo Map Series imagery
collection. We have designed the workflow to handle the following
collections stored in the AWS S3 [TopoReleaseArchive] directory:

   ```
   s3://topographic-upload/TopoReleaseArchive/NZTopo50_GeoTif_Gridless/
   s3://topographic-upload/TopoReleaseArchive/NZTopo250_GeoTif_Gridless/
   ```

To support this workflow, we have introduced new [Topo Raster processes]
to the `basemaps\cogify` package. This workflow executes these new
processes as part of its process flow.

   **This workflow has two key steps:**

   For an NZTopo Map Series imagery collection:

     - Generate a directory structure of STAC Item and Collection files
- Standardise the contained GeoTIFF imagery files and save them within
the generated directory structure

From this, we can pass the resulting directories into the
[imagery-import-cogify] workflow to process the standardised imagery
into basemaps, either for previewing, or publishing.

2. **`imagery-import-cogify` (update)**

This work also updates the [imagery-import-cogify] workflow to support
the new background parameter for the CLI `create` command of the
`basemaps/cogify` package. This allows the command to override any
transparent pixels with a fixed colour. In this work's context, this
allows us to eliminate [imagery artifacts] that used to appear around
the COGs, particularly for NZTopo Map Series imagery.

<!-- external links -->

[NZ Topo Gridless Maps]:
https://basemaps.linz.govt.nz/@-41.8899962,174.0492437,z5?style=topo-raster&i=topo-raster

[as a tile service]:
https://github.com/linz/basemaps/issues/3365#issuecomment-2445444420
[TopoReleaseArchive]:
https://linz-topographic-upload.s3.ap-southeast-2.amazonaws.com/topographic/TopoReleaseArchive/

[Topo Raster processes]: https://github.com/linz/basemaps/pull/3388

[topo-maps-standardising]:
https://github.com/linz/topo-workflows/pull/932/files#diff-985cd50a4d3d95216b2aac800f8f1590a1ac94d4e3d1041ade2f09ce0e1b5bc6
[imagery-import-cogify]:
https://github.com/linz/topo-workflows/pull/932/files#diff-5b88dd6435a818dfac26161a8270f7f4ae499f9771f91e7c8ac08d807f033845

[background parameter]: https://github.com/linz/basemaps/pull/3379
[imagery artifacts]:
https://github.com/linz/basemaps/pull/3379#issue-2734358798