feat: validate STAC of only updated files TDE-1359 (#974)
#### Motivation

When updating an existing Collection, some `links` `item` (the ones not
updated) may not be located in the scratch (or `*/flat/`) directory.
This causes a recursive stac-validation on the Collection to fail (the
system can't locate the linked Items as their path is relative). This PR
takes the approach of validating the Collection not recursively
(excluding its links) and validating every single STAC Items that have
been created and saved in the scratch directory during the process. _The
already published STAC documents are validated on a regular basis._

#### Modification

- Allow `stac-validate-parallel` to validate any file, not only
Collections
- A temporary solution of validating any JSON files has been implemented
as the scratch directory has only STAC document as JSON files. This
could be fixed by allowing the [`stac-validate`
command](https://github.com/linz/argo-tasks/tree/master/src/commands/stac-validate)
to validate a path to a directory and filtering STAC documents before
validation.
- Condition the STAC validation based on the `odr_url` value: if not
empty, that is a re-supply and we should not validate the Collection
recursively but use the other approach.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title