refactor: remove command override for containers that specifiy a entrypoint (#985)
### Motivation

Some containers specify a `ENTRYPOINT` by using "command" we override
the entrypoint, argo-tasks is about to change its `ENTRYPOINT` which
would require this repository to update too, however if we do not
override the `ENTRYPOINT` here we when argo-tasks updates its entrypoint
there will be no changes required here.


### Modifications

Remove all `command` overrides for `argo-tasks` 

### Verification

run a few testing workflows with and without the `command` ovveride.