feat: --delete-all-existing-items TDE-1443 (#997)
### Motivation

https://github.com/linz/topo-imagery/pull/1306 allows to update a
Collection while deleting all its current Items, to be replaced by a new
supply.

### Modifications

- Add a workflow parameter in the `imagery-standardising`
WorkflowTemplate so the user can choose if they want to delete the
existing Items from the Collection when updating an existing dataset.
This flag takes effect only if it's a re-supply, i.e. an `odr_url` has
been provided.

### Verification

Ran a workflow.