ci: deploy service account and semaphore in test namespace (#1026)
### Motivation

The current service account `workflow-runner-sa` lives in the `argo`
namespace only. We need a service account in the test `pr-*` namespace
in order to do operations on the test workflows within the test
namespace. This PR also adds the missing sempahore deployment in the
test namespace.
<!-- TODO: Say why you made your changes. -->

### Modifications

- Create a service account and role binding in the test namespace. 
- Deploy the semaphore

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

Workflows in the test namespace can be ran.
<!-- TODO: Say how you tested your changes. -->