feat: use existing description for merged DEM hillshades TDE-1498 (#1052)
### Motivation

:mega: **Note:** linked to
https://github.com/linz/topo-imagery/pull/1338 which should be merged
and released before this PR is merged.

The DEM merged hillshades need a custom description in the same way that
some datasets need a custom title. Add an option to keep existing
descriptions (load them from the ODR URL).

### Modifications

Added a new flag `--keep-description` when updating the collection for
the merged hillshades.

### Verification

Tested with a workflow. The collection was created with the existing
description.