feat: temporary archive workflow for custom target paths TDE-1605 (#1058)
### Motivation

There is a need to archive data to defined locations so that the
structure of existing staging buckets can be amended when copying across
to archive bucket. This is to create a more organised archive file
structure. This is a temporary workflow before longer term requirements
are assessed.

### Modifications

New workflow based on the archive workflow, but containing a step that
checks the `source` parameter to check it is suitable for archiving
(minimum depth of 2 levels to make sure entire supplier folders or drive
folders or the whole bucket are not requested). Unlike the standard
archive workflow, the target location is not controlled.

### Verification

Tested on nonprod Argo Workflows environment.