feat: add supplied capture area when creating collection TDE-1604 (#1063)
:mega: Do not merge until https://github.com/linz/topo-imagery/pull/1374
is merged and released. :mega:

### Motivation

For 3D Coastal mapping datasets the capture areas automatically
generated when standardising are very complex. This change is to allow a
simpler capture area to be added to a collection rather than calculating
it from the TIFF footprints.

### Modifications

Added a new parameter `supplied_capture_area` to be used by the
`standardise-validate` and `create-collection` steps. The
`standardise-validate` step uses it to override the
`create_capture_area` option and prevent unneccessary generation of
footprint files.

### Verification

Ran workflows manually against topo-imagery PR container.