fix: move delete-empty-folders to copy template TDE-1640 (#1068)
### Motivation

The previous deletion of empty folders relied on an extra copy status
parameter that the archive workflow used to determine whether to delete
empty folders. This parameter had to be used because it's not possible
to access dag task status in a referenced workflow template, and the
deletion of empty folders should not be done if the copy task has been
skipped e.g. if there are no files to archive.

However, there was an unintended consequence that other
workflowtemplates that reference the copy template failed when trying to
resolve the status parameter of the copy.

### Modifications

This moves the deletion of empty folders into the copy template so that
the copy status can be directly accessed. This removes the need for
setting a status parameter explicitly, and resolving that parameter,
however, it does mean that the step is visible in all copy workflows,
not just archive workflows, even if skipped.

### Verification

Manual test on Argo Workflows Non Prod.