feat: unarchive workflow TDE-1553 (#1049)
### Motivation
Archived data stored in S3 Glacier Deep Archive can not be immediately
accessible and requires a restore operation before it can be read.
Manual process to do that is not trivial for most of our users.

### Modifications

This change introduces a workflowTemplate that allows users to easily
initiate a restore for a set of objects under a specified prefix in an
archive bucket. This is the "Part 1" of the unarchive process. A
cronWorkflow will be implemented to verify when the files have been
actually restored and notify (Slack) users.

#### Workflow

- Add a `unarchive` workflowTemplate that manages simple restoration
request (one location, no filters on suffixes).

#### k8s secrets

- Create some secrets to be used by the workflow

#### IAM Roles

- Allow `Workflows-EksWorkflowsArgoRunnerServiceAccountRole` to create
S3 Batch Operations jobs
- Allow `Workflows-EksWorkflowsArgoRunnerServiceAccountRole` to pass the
S3BatchRestoreRole IAM role

### Verification

Test workflows

---------

Co-authored-by: Alice Fage <afage@linz.govt.nz>