feat: verify restore and copy files TDE-1592 (#1071)
### Motivation

As a Data Manager I want to move data from the
`linz-topographic-archive` / `linz-hydrographic-archive` bucket to a
designated bucket so that I can unarchive data when it is required again

### Modifications

This PR contains a cron job that triggers a copy-restore workflow. This
lists outstanding S3 Glacier Batch Restore request manifests (without
the suffix `.done`) and uses the Argo Tasks `verify-restore` command to
check the status of the files referenced in the restore manifests. Once
all the files are restored they are copied to the relevant sharing
bucket.

### Verification

Tested on Argo Workflows nonprod and prod environments.