fix: `check-restore-copy` workflow should mark restore manifest `.done` after copying files succeeded TDE-1703 (#1090)
### Motivation

The restore manifest should be renamed with the `.done` suffix only
after a successful copy of the files to unarchive to their target. At
the moment, this manifest is marked "done" before the copy gets
triggered, avoiding the next job run to retry this copy if the copy has
been unsuccessful the previous time.
See https://github.com/linz/argo-tasks/pull/1286

### Modifications

- Mark the restore manifest to `.done` when the copy is successful
<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

Ran workflows
<!-- TODO: Say how you tested your changes. -->