fix(unarchive): `restore` task should fail when job submission fails TDE-1692 (#1097)
### Motivation

The `restore` task should fail when the AWS Batch job submission fails
rather than logging an error message but continuing the process. We want
the workflow to fail so we know there is an issue.

### Modifications

Throw an error when JOB_ID is empty or null avoiding the rest of the
process to continue.
<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

Test workflow
<!-- TODO: Say how you tested your changes. -->