chore: increase standardising semaphore to 8 (#1108)
### Motivation

Being able to run only 2 standardising workflows at a time is slowing
down processing now more teams are using Argo Workflows. We should be
able to scale further than that. Basemaps imports are up to 10 at a
time.

### Modifications

Increased the `standardising` semaphore to 8, and removed the `bulk`
semaphore as this is no longer required.