docs: add update config on finalize step TDE-1733 (#1114)
### Motivation

Once the EKS cluster is re-installed, its API endpoint changes. The
users need to update their `aws` config in order to access to the new
API endpoint using the command line tools such as `kubectl` or `argo`.

### Modifications

Modify the "Finalise" step to ask to tell users to run the `aws` config
update command.
<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

This command solved the issue for an user after the EKS cluster got
re-installed.
<!-- TODO: Say how you tested your changes. -->